Expose CommandBarDelegate to notify implementer of CommandBar events (#1175)
PROBLEM
There is a need for consumers of the CommandBar to be able to receive notifications about when scroll events occur.

WORK DONE
Implement a CommandBarDelegate which exposes a method, commandBarDidScroll(_ commandBar: CommandBar) that notifies the implementor that a scroll even occurred.

An evaluation of alternative approaches was explored. Alternatives include 1.) Utilizing NSNotification; 2.) Have the consumer use KVM to monitor for scroll events. Both of these seemed less than ideal so I went with this protocol approach.

VERIFICATION
- Verified static library compiles
- Updated the CommandBarDemoController with a switch to turn on animations when the CommandBarDelegate events occur. The animation simply changes the CommandBar background color to blue and then back to the original background color.