Fix RTL behavior for CommandBar Pinned Items (#1236)
It was discovered that the RTL behavior for the pinned items in CommandBar was off. The leading and trailing groups were switching sides as expected, but the additional transform being applied `configureHierarchy` is causing the views inside the group to be shown in LTR order.

Fix: Remove the transform for the leading and trailing groups inside `configureHierarchy` and let the default behavior from the UIStackView handle the transition.
(cherry picked from commit 8273727afd7902fbe4054ddaa6c0a10c3419a7e5)