make sure headerfooterview isn't clear background and actually listen to alias token (#1570)
If the tableviewStyle is .plain, because the headerfooterview by default is .clear color, when the cell scrolls, it draws in the background. Even though TableViewHeaderFooterView isn't fluent2 tokenize yet, lets use the alias token to update the background color, if the client configures what TableViewStyle this headerfooter view is hosted in.