bug fix: unreaddot interfering tableviewcell accessibilityLabel (#1618) (#1619)
When #1446 introduced unread dot to tableviewcell, it started to customize accessibilityLabel of the cell. Couple of issues observed
(1) when tableview scrolls, and it dequeues an existing cell identifier, it will still contain the old accessibilityLabel value and not grab the latest title of the cell. We assume Apple behind the scene groups all the UILabel inside the cell and aggregate the cell's accessibilitylabel correctly. But, once you custom call accessibilityLabel and set it yourself, this hidden feature no longer works.

(2) unread dot assumes that the titleLabel.title is the right string. When in fact, it could be titleLabel.attributedString that has the right value for the cell.

We want to fix this bug quickly with minimal impact of loc and code changes. For now, similar to how we handle boolean cell (which is cell with uiswitch) move the "unread" string to AccessibilityValue instead. Currently our loc string is "%@, unread" which isn't ideal, but VoiceOver seems to just read out fine. We will need to refactor a bit later to get this corrected.