use sizeToFit rather than layoutifneeded in cell (#1648)
When we try to use the static function to figure out the cell's height, we try to take customAccessoryView size into account. However, #612 calls layoutIfNeeded which isn't entirely correct because the accessoryView might not be layout at this point. Rather, lets call sizeToFit() which tried to get the intrinsic contentsize of the accessoryView to calculate.

miscellaneous:
remove all the workaround when we were adding badge in the tableviewcell in the demo.