use ios 17 api (#2057)
iOS 17 added new accessibility trait which custom control that now notify to users it is a toggle. BooleanCell is a custom UITableViewCell which we had our own accessibility strings to mimic OS behavior. We can eventually remove this custom string if all of our cells with uiswitch contains .toggleButton trait.